FetchTelegram/requirements.txt
Telethon==1.38.1
uvicorn==0.34.0
fastapi==0.115.8
python-dotenv==1.0.1